Rename `src` directory
