+++ isr/data/extract_kaldi_xvectors.sh
